time series task/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
joblib==1.4.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
patsy==0.5.6
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.4.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
statsmodels==0.14.4
threadpoolctl==3.5.0
tzdata==2024.2
xgboost==2.1.1
